backend/requirements.txt
certifi==2020.11.8
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2
django-rest-swagger==2.2.0
djangorestframework-jwt==1.11.0
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
openapi-codec==1.3.2
Pillow==8.0.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
PyJWT==1.7.1
pytz==2020.4
requests==2.25.0
simplejson==3.17.2
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
